requirements.txt
pyyaml>=6.0.1
pandas>=2.0.0
requests>=2.31.0
python-dotenv>=1.0.0
pydantic>=2.0.0
openai~=1.75.0
setuptools~=58.0.4
jsonpath-ng>=1.7.0
networkx>=3.0.0
jinja2>=3.0.0
